pyproject.toml
[tool.poetry]
name = "repl_nix_django-calendarapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-auth-oauthlib = "^1.0.0"
xds-protos = "^0.0.11"
Django = "^4.2.1"
google-api-python-client = "^2.87.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
